Reorganise the page into a component
